🦄 refactor(user): 重构用户 DTO 结构，优化包路径和类组织
删除了冗余的 DTO 类，并将相关 DTO 重新组织到更合适的包路径中，以提高代码可读性和可维护性。
