+++ backend/blog-backend/src/main/java/com/kisesaki/blog/user/dto/follow/UserFollowDto.java
@@ -1,4 +1,4 @@
-package com.kisesaki.blog.user.dto;
+package com.kisesaki.blog.user.dto.follow;
 
 import java.time.OffsetDateTime;
