+++ backend/blog-backend/src/main/java/com/kisesaki/blog/user/service/AdminUserService.java
@@ -1,0 +1,5 @@
+package com.kisesaki.blog.user.service;
+
+public class AdminUserService {
+    
+}
